# internal/web/web.go
// Package web serves the embedded Paperwing single-page application.
package web

import (
	"embed"
	"io/fs"
	"net/http"
	"path"
	"strings"
)

// dist is generated by `npm run build` in the frontend directory.
//
//go:embed all:dist
var bundled embed.FS

// New combines the JSON API with the embedded frontend. Unknown GET routes
// fall back to index.html so client-side navigation keeps working.
func New(api http.Handler) http.Handler {
	dist, err := fs.Sub(bundled, "dist")
	if err != nil {
		panic(err)
	}
	files := http.FileServer(http.FS(dist))

	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if isAPIPath(r.URL.Path) || (r.Method != http.MethodGet && r.Method != http.MethodHead) {
			api.ServeHTTP(w, r)
			return
		}

		cleanPath := strings.TrimPrefix(path.Clean(r.URL.Path), "/")
		if cleanPath != "." {
			if info, statErr := fs.Stat(dist, cleanPath); statErr == nil && !info.IsDir() {
				if strings.HasPrefix(cleanPath, "assets/") {
					w.Header().Set("Cache-Control", "public, max-age=31536000, immutable")
				}
				setSecurityHeaders(w)
				files.ServeHTTP(w, r)
				return
			}
		}

		setSecurityHeaders(w)
		w.Header().Set("Cache-Control", "no-cache")
		clone := r.Clone(r.Context())
		clone.URL.Path = "/"
		files.ServeHTTP(w, clone)
	})
}

func isAPIPath(value string) bool {
	return value == "/healthz" ||
		value == "/auth" || strings.HasPrefix(value, "/auth/") ||
		value == "/api-tokens" || strings.HasPrefix(value, "/api-tokens/") ||
		value == "/accounts" || strings.HasPrefix(value, "/accounts/") ||
		value == "/emails" || strings.HasPrefix(value, "/emails/")
}

func setSecurityHeaders(w http.ResponseWriter) {
	w.Header().Set("X-Content-Type-Options", "nosniff")
	w.Header().Set("Referrer-Policy", "no-referrer")
	w.Header().Set("X-Frame-Options", "SAMEORIGIN")
	w.Header().Set("Content-Security-Policy", "default-src 'self'; img-src 'self' data:; style-src 'self' 'unsafe-inline'; script-src 'self'; connect-src 'self'; frame-src 'self' data:; base-uri 'none'; form-action 'self'")
}